+++ src/entry/content/index.js
@@ -9,11 +9,14 @@ import { Provider } from 'react-redux';
 import store from '../../redux/store';
 
 import { initMultipleTranslate, initSingleTranslate } from '../../redux/init';
-import { initOptions, listenOptionsChange } from '../../public/options';
+import { initOptions } from '../../public/options';
 import { getLocalStorage, onExtensionMessage } from '../../public/chrome-call';
 import defaultOptions from '../../constants/defaultOptions';
 
 import '../../styles/global.css';
+
+// inject style
+import '../../public/inject-style';
 
 const init = (options) => {
   initOptions(options);
@@ -22,11 +25,6 @@ const init = (options) => {
 
   const app = document.createElement('div');
   app.id = 'sc-translator-root';
-  app.className = options.darkMode ? 'dark' : 'light';
-
-  listenOptionsChange(['darkMode'], (opts) => {
-    app.className = opts.darkMode ? 'dark' : 'light';
-  });
 
   document.body.appendChild(app);
   ReactDOM.render(