+++ src/entry/options/index.js
@@ -4,6 +4,9 @@ import Options from '../../components/Options';
 import './style.css';
 import '../../styles/global.css';
 
+// inject style
+import '../../public/inject-style';
+
 document.body.id = 'sc-translator-root';
 
 ReactDOM.render(